done working on the signin signup and the category crud started working on the posts crud

+++ pages/index.php
@@ -1,22 +1,25 @@
 <?php
 include '../classes/user.class.php';
 include '../classes/dashboard.php';
+
+if(isset($_GET['action']) && $_GET['action'] === 'logout') {
+    $spectateur = new Spectateur();
+    $spectateur->logout();
+}
+isset($_SESSION['name']) ?: header('location:../pages/signin.php');
 $category = new Categories();
 $category->addCategory();
 $getcategory = new Database();
 $getcategory->getCategory('categories','*');
 $result=$getcategory->sql;
-$id = $_POST['id'];
-$getcategory->delete('categories',"id='$id'");
-// $categories= new Categories();
-// $categories->getCategories();
-// if(isset($_GET['action']) && $_GET['action'] === 'logout') logOut();
-
-// function logOut(){echo 'lahoula';
-//     $spectateur = new Spectateur();
-//     $spectateur->logout();
-//     header('location:../pages/signin.php');
-// }
+$rows=[];
+while($resultt=$result->fetch(PDO::FETCH_ASSOC)){
+    array_push($rows,$resultt);
+};
+if(isset($_GET['delete'])){
+    $id = $_GET['delete'];
+    $getcategory->delete('categories',"id='$id'");
+}
 ?>
 
 <!DOCTYPE html>
@@ -48,16 +51,16 @@ $getcategory->delete('categories',"id='$id'");
                         <img src="../assets/img/ppl.png" class="pp">
                         <button type="button" class="btn dropdown-toggle text-light" data-bs-toggle="dropdown" aria-expanded="false">
                         <?php
-                            // if(isset($_SESSION["UserName"]))
-                            // echo $_SESSION["UserName"];
+                            if(isset($_SESSION["name"]))
+                            echo $_SESSION["name"];
                         ?>
                         </button>
                         <ul class="dropdown-menu">
                             <li><button class="dropdown-item" type="button">Settings</button></li>
                             <li><hr class="dropdown-divider"></li>
-                            <li><form method="POST">
-                                <button class="dropdown-item" type="submit" name="logout" href="../classes/user.class.php?&action=logout">Logout</button>
-                                </form>
+                            <li><a class="link" href="index.php?action=logout">
+                                    <button class="dropdown-item">Logout</button>
+                                </a>
                             </li>
                         </ul>
                     </div>
@@ -82,7 +85,7 @@ $getcategory->delete('categories',"id='$id'");
                         <div class="stats col-3 text-light ms-2">
                             <h4 class="mt-2">Categories</h4>
                             <h4 class="mt-3"><?php 
-                            // counterUser() ?>54</h4>
+                            echo count($rows); ?></h4>
                         </div>
                     </div>
                     <div class="user ms-4 col-4 row h-50 mt-2" id="post">
@@ -132,8 +135,9 @@ $getcategory->delete('categories',"id='$id'");
                     <th class="col-1"></th>
                     <th class="col-2">Admin</th>
                     <th class="col-2">Date & Time</th>
-                    <th class="col-4">Articles</th>
-                    <th class="col-2">Category</th>
+                    <th class="col-2">Title</th>
+                    <th class="col-3">Articles</th>
+                    <th class="col-1">Category</th>
                     <th class="col-1"></th>
                 </tr>
             </thead>
@@ -154,6 +158,10 @@ $getcategory->delete('categories',"id='$id'");
                         <div class="modal-body">
                             <input type="hidden" id="task-id" name = 'id'>
                             <div class="mb-3">
+                                <label class="col-form-label">Title:</label>
+                                <input type="text" class="form-control" aria-label="Sizing example input" aria-describedby="inputGroup-sizing-sm">
+                            </div>
+                            <div class="mb-3">
                                 <label for="message-text" class="col-form-label">Post:</label>
                                 <textarea class="form-control" id="message-text"></textarea>
                             </div>
@@ -161,12 +169,9 @@ $getcategory->delete('categories',"id='$id'");
                                     <label class="form-label">Category</label name priority>
                                     <select class="form-select" id="task-priority" name="category">
                                         <option value="">Please select</option>
-                                        <option value="1">Game controller</option>
-                                        <option value="2">Memory unit</option>
-                                        <option value="3">Audio/Video cable</option>
-                                        <option value="4">Case</option>
-                                        <option value="5">PC</option>
-                                        <option value="6">Software accessorie</option>
+                                    <?php  for($i=0;$i<count($rows);$i++) {?>
+                                        <option value="<?= $rows[$i]['id']?>"><?php echo $rows[$i]['categ'];?></option>
+                                        <?php } ?>
                                     </select>
                                 </div>
                             <div class="mb-3">
@@ -185,21 +190,21 @@ $getcategory->delete('categories',"id='$id'");
     </div>
     <div id="categories">
         <div class="d-flex justify-content-end mt-4">
-            <a href="#modal-task-category" data-bs-toggle="modal" class="add-btn btn btn-rounded rounded-pill" onclick="document.getElementById('form').reset()"><i class="bi bi-plus me-2 ms-n2 text-success-900"></i> Add Category</a>
+            <a href="#modal-category" data-bs-toggle="modal" class="add-btn btn btn-rounded rounded-pill" onclick="document.getElementById('form').reset()"><i class="bi bi-plus me-2 ms-n2 text-success-900"></i> Add Category</a>
         </div>
         <div class="row mt-4 ms-5 fs-5">
-            <?php  while ($row = $result->fetch(PDO::FETCH_ASSOC)) {?>
+            <?php  for($i=0;$i<count($rows);$i++) {?>
             <div class="category col-2 p-3 mx-auto">
                 
-                <div class="d-flex justify-content-end"><a data-bs-toggle="modal" data-id="<?php echo $row['id']; ?>" href="#deleteModal" id="del" class="btn btn-danger btn-sm"><i class="bi bi-x"></i></a></div>
+                <div class="d-flex justify-content-end"><button data-bs-toggle="modal" data-bs-target="#deleteModal" data-id="<?php echo $rows[$i]['id']; ?>"  id="del" class="btn btn-danger btn-sm" onclick="deleteCategory(<?= $rows[$i]['id']?>)"><i class="bi bi-x"></i></button></div>
                 
-                <div><?php echo $row['categ'];?></div>
+                <div><?php echo $rows[$i]['categ'];?></div>
             </div>
         <?php } ?></div>
-        <div class="modal fade" id="modal-task-category">
+        <div class="modal fade" id="modal-category">
             <div class="modal-dialog">
                 <div class="modal-content">
-                    <form method="POST" id="form" enctype="multipart/form-data">
+                    <form method="POST" id="form">
                         <div class="modal-header">
                             <h5 class="modal-title">Add Category</h5>
                             <a href="#" class="btn-close" data-bs-dismiss="modal"></a>
@@ -212,29 +217,29 @@ $getcategory->delete('categories',"id='$id'");
                         </div>
                         <div class="modal-footer">
                             <a href="#" class="btn btn-white" data-bs-dismiss="modal">Cancel</a>
-                            <button type="submit" name="save" class="btn btn-primary task-action-btn ms-1" id="task-save-btn">Save</button>
+                            <button type="submit" name="submit" class="btn btn-primary task-action-btn ms-1" id="task-save-btn">Save</button>
                         </div>
                     </form>
                 </div>
             </div>
         </div>
-<div class="modal fade" id="deleteModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
-  <div class="modal-dialog">
-    <div class="modal-content">
-      <div class="modal-header">
-        <h1 class="modal-title fs-5" id="exampleModalLabel">HoldOOON!!!</h1>
-        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-      </div>
-      <div class="modal-body">
-        Do you really want to delete this category?
-      </div>
-      <div class="modal-footer d-flex justify-content-between">
-        <button type="button" class="btn btn-secondary ms-5" data-bs-dismiss="modal">NO</button>
-        <button type="button" class="btn btn-primary">ofc</button>
-      </div>
-    </div>
-  </div>
-</div>
+        <div class="modal fade" id="deleteModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
+            <div class="modal-dialog">
+                <div class="modal-content">
+                <div class="modal-header">
+                    <h1 class="modal-title fs-5" id="exampleModalLabel">HoldOOON!!!</h1>
+                    <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                </div>
+                <div class="modal-body">
+                    Do you really want to delete this category?
+                </div>
+                <div class="modal-footer d-flex justify-content-between">
+                    <button type="button" class="btn btn-secondary ms-5" data-bs-dismiss="modal">NO</button>
+                    <button id="deleteButton" type="button" class="btn btn-primary" name="deletecat" data-bs-dismiss="modal">ofc</button>
+                </div>
+                </div>
+            </div>
+        </div>
     </div>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-OERcA2EqjJCMA+/3y+gxIOqMEjwtxJY7qPCqsdltbNJuaOe923+mo//f6V8Qbsw3" crossorigin="anonymous"></script>
     <script src="../assets/scripts/scripts.js"></script>
